+++ pwm.h
@@ -2,7 +2,7 @@
 #define PWM_H_
 
 void SetMotorDutyCycle(unsigned int DutyCycle, int motorLeft, int dir);
-void SetServoDutyCycle(unsigned int DutyCycle);
+void SetServoDutyCycle(double DutyCycle);
 void InitPWM();
 void PWM_ISR();
